test: isolate contacts and mail stores from Electron bootstrap
